break task-list up into task-families and task-revisions
